Create Analyze Consumer Buying Behavior.ipynb
You have been tasked with analyzing shopping data for a web site! The data is in JSON format and available with this notebook.
On the website, each user logs in using their personal account and can purchase products as they browse the list of products offered. Each product has a sales value. Age and gender data for each user has been collected and is provided in the JSON file.
Your job is to deliver an analysis of consumers' buying behavior. This is a type of common activity carried out by Data Scientists and the result of this work can be used, for example, to feed a Machine Learning model and make predictions about future behaviors.
But on this mission, you will analyze consumers' buying behavior using the Python language Pandas package and your final report should include each of the following items: